Functional decryption/encryption
The Crypto.Copy_Completed_Data implementation did not implement the flow
of data for plain and cipher text completely. Hence, the actual
encryption and decryption of the CBE was not functional until now.

This patch splits the Crypto.Copy_Completed_Data into two procedures
Copy_Decrypted_Data (for read operations) and Copy_Encrypted_Data (for
write operations) to clear up the flow and buffering of information.

Thanks a lot to @ssumpf for his great support!