cbe: store current query VBA out-side
Resetting 'Curr_Query_Prim' by assigning 'Invalid_Primitive' also sets
the Block_Number back to 0. The invalid object is then used to
calculate the next VBA, which in this case will always be 64. This
will lead to an endless loop when the FT has to query more than 2
branches to find free blocks.