cbe: snapshot interface
Adds sub-programs to the CBE library interface that can be used to manage
snapshots explicitely and adapts the client interface of the CBE library in a
way that one can read from any specific snapshot (the current snapshot can
still be written additionally). The interfaces for sealing generations and
securing superblocks were removed from the CBE library as the new snapshot
interface renders them useless.

New interface to manage snapshots
---------------------------------

:Library.Create_Snapshot:

Start creating a normal or a quarantine snapshot. Returns ID of new snapshot.

:Library.Snapshot_Creation_Complete:

Returns whether the snapshot behind a given ID was the last that was created
and whether the creation finished successful.

:Library.Active_Snapshot_IDs:

Returns an array of IDs of all existing snapshots that can be accessed
currently.

Changes to the existing client interface
----------------------------------------

:Library.Submit_Client_Request:

Received an additional "in" parameter that denotes the ID of the snapshot the
request refers to. A value of 0 has the special meaning that the request
refers to the current snapshot.

:Library.Is_Securing_Superblock:
:Library.Is_Sealing_Generation:
:Library.Start_Securing_Superblock:
:Library.Start_Sealing_Generation:

Were removed.