cbe: gnatprove in mode "check_all" succeeds
This required especially the adaption of sub-program interfaces according to
the requirements of SPARK (including some sub-programs of the CBE library
interface - see below) and the separation of debugging code into a dedicated
package with "SPARK_Mode => Off". To be able to have calls to the debugging
package in SPARK code without breaking proofs, an  ADC file was added
containing a global static switch that, by default, turns the evaluation of
code marked with "pragma Debug" off.

Changes to the CBE library interface
------------------------------------

:Supply_Client_Data:

Was a function, is now a procedure that provides its former return value
through a new "out" parameter.

:Drop_Completed_Request:

Removed request "in" parameter as the current implementation doesn't use it.