lib/cbe: secure superblock on snapshot creation
Whenever a branch is changed we also have to update the superblock so
the list of active snapshots stays consistent. Not updating the super-
block was due to the fact the that the securing of the superblock is
actually a different operation than syncing, e.g. updating, the super-
block.