cbe: clean up client interface
All sub-programs of the client interface of the CBE library now contain the
phrase "client" to make distinguishing the different CBE interfaces easier.
Also, the implementation of the sub-program for dropping completed client
requests was adjusted in the way that it now drops a specific request and not
any completed request as before.

Changes to the CBE library interface
------------------------------------

:Client_Request_Acceptable:

Renamed, former name was Request_Acceptable.

:Submit_Client_Request:

Renamed, former name was Submit_Request.

:Peek_Completed_Client_Request:

Renamed, former name was Peek_Completed_Request.

:Drop_Completed_Client_Request:

Renamed, former name was Drop_Completed_Request. Now receives the request that
shall be dropped as additional "in" parameter.

:Client_Data_Index:

Renamed, former name was Give_Data_Index.