lib/cbe: use root PBA and hash for SN creation
This basic idea is to check last secured generation and root hash.
Since the working snapshot gen is last secured + 1 a sync will make
it persistent, just make sure that we really create the SN only when
necessary, e.g. the hash has changed.

(Rename the securing part to sync superblock b/c that's what it
really is and use securing for when we give the TA the hash of the
current SB is still pending.)