lib/cbe: get rid of avoidable mod types
A lot of types in the CBE were still modular types eventhough it was not
necessary. Values of these types can wrap and the ranges are often much bigger
than required. Unused parts of such ranges are misused for mixing in
additional meanings. Furthermore, proving code is harder with modular types
and their flexibility. This commit replaces, as far as possible, all mod types
with types of a domain specific range. As a side precondition, values with
special meanings are replaced with additional variables of different type and
static constraints of the CBE are formalized.