lib/cbe: hardcode layout of metadata blocks
The free-number member in the superblock type was written to a bad offset
inside the block when converting superblocks to data blocks. This was because
we used 'Size attributes to calculate the offset. One of the types we used with
'Size was changed over time from a storage to a normal range type and therefore
changed in size. The layout of metadata blocks should be hardcoded and not
depend on implementation details of components that use it.