cbe: store Superblock_ID in superblock
The 'Superblock_ID' is incremented by one very time a superblock is
written to disk. The ID of the initial superblock is always 1.

(The intention is to identify the most recent superblock even when
no snapshot creation took place, i.e. only 'SYNC' block request
were used to store changes on disk. This may change in the future
when the TA is securing the written SB and can select the SB by
checksum.)