lib/cbe: store level and LRU counter in cache
Whenever a cache entry is accessed via the 'Cache.Data_Index' procedure
the level within tree to which the entry belongs is passed in. This is
used to check that the entry is indeed used for the proper level. A
cache entry must never by used by different levels w/o prior eviction.