lib/cbe: check also hashes of leaf nodes
When the VBD finishes an address translation he has access to the hash of the
determined leaf node. In case it's a translation for a read request, we let the
VBD hand out the hash to us so we can store the hash in the Block_IO until the
Block_IO has finished reading the block data from the device. Then the Library
module is able to check the read data against the hash before it enters the
cache. This is somehow a quickfix as it stands to reason whether the Library
module is the place where to check hashes and whether the Block_IO is the place
where to store hashes.