cbe_block: don't initialize leaf nodes
The generation and hash initially are both set to zero for leaf nodes while
their data is not initialized anymore. The idea is that the CBE must not read
their content but assume that it is all zeros when it reads generation 0. This
spares a significant overhead when initializing large virtual block devices.